// src/lib/api/schema.ts
export interface components {
    schemas: {
// ...
        IssueDetail: {
            readonly id: number;
            subject: string;
            description: string;
            readonly creator: components["schemas"]["UserMini"];
            readonly assigned_to: components["schemas"]["UserMini"];
            readonly type: string;
            readonly severity: string;
            readonly priority: string;
            readonly status: string;
            readonly tags: string;
            /** Format: date */
            deadline?: string | null;
            due_date_reason?: string | null;
            /** Format: date-time */
            readonly created_at: string;
            /** Format: date-time */
            readonly modified_at: string;
            readonly watchers: components["schemas"]["UserMini"][];
            readonly comments: components["schemas"]["CommentRead"][];
            readonly attachments: components["schemas"]["AttachmentRead"][];
            readonly activity: components["schemas"]["ActivityRead"][];
        };
// ...
        UserMini: {
            readonly id: number;
            /** @description Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only. */
            username: string;
            /**
             * Profile picture
             * Format: uri
             */
            avatar?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}